+++ src/adapter/types.ts
@@ -1,7 +1,7 @@
 import type { StoreEventsAbiItem, StoreEventsAbi } from "@latticexyz/store";
 import type { UnionPick } from "@latticexyz/common/type-utils";
+import type { BlockLogs } from "@primodiumxyz/sync-stack/types";
 import type { Log } from "viem";
-import { BlockLogs } from "@primodiumxyz/sync-stack/types";
 
 type StoreEventsLog = Log<bigint, number, false, StoreEventsAbiItem, true, StoreEventsAbi>;
 export type StorageAdapterLog = Partial<StoreEventsLog> & UnionPick<StoreEventsLog, "address" | "eventName" | "args">;
