+++ src/lib/external/mud/queries.ts
@@ -16,9 +16,11 @@ import { observable, ObservableSet } from "mobx";
 import isEqual from "fast-deep-equal";
 import { useEffect, useMemo, useState } from "react";
 
-import type { TableUpdate } from "@/queries";
-import type { BaseTable, Properties } from "@/tables";
-import { type Entity, type Schema, tableOperations, Type } from "@/lib";
+import type { TableUpdate } from "@/queries/types";
+import type { BaseTable, Properties } from "@/tables/types";
+import type { Entity } from "@/lib/external/mud/entity";
+import { type Schema, Type } from "@/lib/external/mud/schema";
+import { tableOperations } from "@/lib/external/mud/tables";
 const {
   hasEntity,
   getEntityProperties,
