+++ src/lib/external/mud/world.ts
@@ -1,12 +1,7 @@
-import type { BaseTable } from "@/tables";
-import {
-  getEntityHex,
-  getEntitySymbol,
-  tableOperations,
-  transformIterator,
-  type Entity,
-  type EntitySymbol,
-} from "@/lib";
+import type { BaseTable } from "@/tables/types";
+import { getEntityHex, getEntitySymbol, type Entity, type EntitySymbol } from "@/lib/external/mud/entity";
+import { tableOperations } from "@/lib/external/mud/tables";
+import { transformIterator } from "@/lib/external/mud/utils";
 const { hasEntity: tableHasEntity, removeEntity: tableRemoveEntity } = tableOperations;
 
 /**
