+++ src/queries/types.ts
@@ -1,5 +1,7 @@
-import type { BaseTable, BaseTableMetadata, Properties, Table } from "@/tables";
-import type { Entity, Schema, World } from "@/lib";
+import type { BaseTable, BaseTableMetadata, Properties, Table } from "@/tables/types";
+import type { Entity } from "@/lib/external/mud/entity";
+import type { Schema } from "@/lib/external/mud/schema";
+import type { World } from "@/lib/external/mud/world";
 
 /* --------------------------------- GLOBAL --------------------------------- */
 /**
